remove rounds
a node would only send new requests after receiving answers to all
previous requests, so it worked in "rounds"
but why not send a request to a new peer as soon as we learn about the
new peer? We don't have to wait for previous responses
also, when a peer has responded, we consider that response correct; we
don't have to wait for another response, at least for safety. It may be
necessary to get a new response from the same peer after it became a
leader for liveness (we eventually find a leader), but we cannot rely on
it for safety. Hence setting Responded to {} was unnecessary.