get rid of Request phase
The condition:
```
    /\ \E m \in Msgs:
        /\ m.type = "Request"
        /\ m.dst = a
        /\ ~ \E mm \in Msgs:
            /\ mm.type = "Response"
            /\ mm.src = a
            /\ mm.dst = m.src
```
implies the condition:
```
    /\ \E src \in Node:
        /\ a \in Peers[src]
        /\ ~ \E mm \in Msgs:
            /\ mm.type = "Response"
            /\ mm.src = a
            /\ mm.dst = src
```

Indeed, if m \in Msgs satisfies the first condition, then a \in
Peers[src] where src = m.src:
it was true when the message was sent (in Request) and it
remains true since the Peers sets only grow over time.

Thus we abstract away the sending of the Request. Instead,
Respond(a) can cheat and inspect the source's Peers directly,
pretending that it received a message from the source. The previous spec
is a refinement of this abstracted spec.