remove Seeds
The Seeds variable was non-deterministically chosen in the initial
state, then it remained constant. It was only used to determine the
initial choice of Peers.

Therefore, if we had two states that had all variables equal
(specifically the Peers variable) except for seeds - those states would
be considered different by TLC, so TLC would have to explore them both
during model checking, even if "logically" the states were equivalent.

Thus we remove Seeds and instead we choose the initial value of Peers
nondeterministically in the initial state.